Add profile settings page and update profile view styles
