+++ public/styles/profile.css
@@ -2,80 +2,60 @@ h1 {
     text-align: center;
 }
 
+.img_profile {
+    aspect-ratio: 1 / 1;
+    object-fit: cover;
+    border-radius: 100%;
+    min-height: 100px;
+    max-height: 150px;
+}
+
 h2 {
-    font-size: 1.2rem;
-}
-
-section {
-    margin-top: 50px;
-    width: 50%;
-    align-self: center;
-    justify-self: center;
-}
-
-.div-structure {
-    display: grid;
-    align-items: end;
-    box-shadow: 0 0px 2px white;
-    border-radius: 10px;
-    padding: 20px;
-    margin-top: 20px;
-    margin-bottom: 20px;
-}
-
-label, .btn-form, input, input[type="file"]::file-selector-button {
-    font-size: 0.7rem;
-}
-
-input[type="text"], input[type="email"], input[type="password"], .btn-form {
-    height: 35px;
-}
-
-.btn-form {
     text-align: center;
-    line-height: 0;
+    font-size: 1.5rem;
 }
 
-label {
-    font-size: 0.8rem;
-    font-weight: bold;
+.section_profile {
+    text-align: center;
+    margin: 20px;
 }
 
-p {
-    font-size: 0.6rem;
+.section_recomendations {
+    width: 90%;
+    display: grid;
+    grid-template-columns: repeat(4, calc(25% - 15px));
+    justify-self: center;
+    align-self: center;
+    gap: 20px;
 }
 
-.div-1 {
-    display: flex;
-    flex-direction: row;
-    gap: 10px;
-    margin-bottom: 0;
+.section_recomendations article {
+    height: 500px;
+    padding: 25px;
 }
 
-.div-1 input[type="text"], .div-1 input[type="email"] {
-    width: 250px;
-    margin-bottom: 0;
-}
-
-.div-1 .btn-form {
-    max-width: 125px;
-    margin-bottom: 0;
+.img_publication {
+    min-height: 100px;
+    max-height: 200px;
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
 }
 
 @media (max-width: 1500px) {
-    section {
-        width: 60%;
-    }
-}
-
-@media (max-width: 1200px) {
-    section {
-        width: 70%;
+    .section_recomendations {
+        grid-template-columns: repeat(3, calc(33% - 10px));
     }
 }
 
 @media (max-width: 1000px) {
-    section {
-        width: 80%;
+    .section_recomendations {
+        grid-template-columns: repeat(2, calc(50% - 10px));
+    }
+}
+
+@media (max-width: 700px) {
+    .section_recomendations {
+        grid-template-columns: repeat(1, calc(100%));
     }
 }
