+++ public/styles/profile_settings.css
@@ -1,0 +1,81 @@
+h1 {
+    text-align: center;
+}
+
+h2 {
+    font-size: 1.2rem;
+}
+
+section {
+    margin-top: 50px;
+    width: 50%;
+    align-self: center;
+    justify-self: center;
+}
+
+.div-structure {
+    display: grid;
+    align-items: end;
+    box-shadow: 0 0px 2px white;
+    border-radius: 10px;
+    padding: 20px;
+    margin-top: 20px;
+    margin-bottom: 20px;
+}
+
+label, .btn-form, input, input[type="file"]::file-selector-button {
+    font-size: 0.7rem;
+}
+
+input[type="text"], input[type="email"], input[type="password"], .btn-form {
+    height: 35px;
+}
+
+.btn-form {
+    text-align: center;
+    line-height: 0;
+}
+
+label {
+    font-size: 0.8rem;
+    font-weight: bold;
+}
+
+p {
+    font-size: 0.6rem;
+}
+
+.div-1 {
+    display: flex;
+    flex-direction: row;
+    gap: 10px;
+    margin-bottom: 0;
+}
+
+.div-1 input[type="text"], .div-1 input[type="email"] {
+    width: 250px;
+    margin-bottom: 0;
+}
+
+.div-1 .btn-form {
+    max-width: 125px;
+    margin-bottom: 0;
+}
+
+@media (max-width: 1500px) {
+    section {
+        width: 60%;
+    }
+}
+
+@media (max-width: 1200px) {
+    section {
+        width: 70%;
+    }
+}
+
+@media (max-width: 1000px) {
+    section {
+        width: 80%;
+    }
+}
