+++ public/styles/profile_setup.css
@@ -26,6 +26,7 @@ img {
     aspect-ratio: 1 / 1;
     object-fit: cover;
     border-radius: 100%;
+    min-height: 150px;
     max-height: 250px;
 }
