+++ src/infraestructure/routes/web.php
@@ -6,6 +6,7 @@ use App\Controller\HomeController;
 use App\Controller\LoginController;
 use App\Controller\RegisterController;
 use App\Controller\ProfileController;
+use App\Controller\ProfileSettingsController;
 
 Router::get("/", [HomeController::class, "index"]);
 
@@ -16,8 +17,9 @@ Router::get("/register", [RegisterController::class, "index"]);
 Router::post("/register", [RegisterController::class, "getData"]);
 
 Router::get("/profile", [ProfileController::class, "index"]);
-Router::post("/profile", [ProfileController::class, "getData"]);
-
 Router::post("/profile/setup", [ProfileController::class, "userSetup"]);
+
+Router::get("/profile/settings", [ProfileSettingsController::class, "index"]);
+Router::post("/profile/settings", [ProfileSettingsController::class, "getData"]);
 
 ?>
