+++ src/view/profile.php
@@ -4,73 +4,43 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Profile</title>
+    <base href="/LoginPHP/public/">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@picocss/pico@2/css/pico.min.css">
     <link rel="stylesheet" href="./styles/navigation.css">
     <link rel="stylesheet" href="./styles/profile.css">
 </head>
 <body>
-    <h1>Bienvenido a tu perfil ...</h1>
+    <h1>Bienvenido a tu perfil</h1>
     <nav>
         <ul>
-            <li><a href="./"><button>Home</button></a></li>
+            <li><a href=""><button>Home</button></a></li>
             <li><a href="login"><button>Login</button></a></li>
             <li><a href="register"><button>Register</button></a></li>
             <li><a href="profile"><button>Profile</button></a></li>
+            <li><a href="profile/settings"><button>Profile/Settings</button></a></li>
         </ul>
     </nav>
-    <section>
-        <form method="POST" action="profile">
-            <label for="image">Foto de perfil</label>
-            <input name="image" type="file">
-            <button class="btn-form" type="submit">Aplicar</button>
-            <input type="hidden" name="action" value="changePhoto">
-        </form>
-
-        <form method="POST" action="profile" class="div-structure">
-            <label for="username">Username</label>
-            <p>Cambia tu nombre de usuario y pulsa <strong>renombrar</strong> para cambiarlo.</p>
-            <div class="div-1">
-                <input name="username" type="text" placeholder="Username">
-                <button class="btn-form" type="submit">Renombrar</button>
-            </div>
-            <input type="hidden" name="action" value="changeUsername">
-        </form>
-
-        <form method="POST" action="profile" class="div-structure">
-            <label for="email">Email</label>
-            <p>Introduce tu correo electronico y pulsa en <strong>verificar</strong>. Una vez hecho te llegara un correo electronico para la verificacion.</p>
-            <div class="div-1">
-                <input name="email" type="email" placeholder="Email">
-                <button class="btn-form" type="submit">Verificar</button>
-            </div>
-            <input type="hidden" name="action" value="verifyEmail">
-        </form>
-
-        <form method="POST" action="profile" class="div-structure">
-            <label for="email">Email</label>
-            <div>
-                <input name="email" type="email" placeholder="Email">
-            </div>
-            <label for="newEmail">Nuevo Email</label>
-            <div>
-                <input name="newEmail" type="email" placeholder="Nuevo Email">
-            </div>
-            <button class="btn-form" type="submit">Cambiar</button>
-            <input type="hidden" name="action" value="changeEmail">
-        </form>
-
-        <form method="POST" action="profile" class="div-structure">
-            <label for="password">Password</label>
-            <div>
-                <input name="password" type="password" minlength="8" placeholder="Password">
-            </div>
-            <label for="newPassword">Nueva Password</label>
-            <div>
-                <input name="newPassword" type="password" minlength="8" placeholder="Nueva Password">
-            </div>
-            <button class="btn-form" type="submit">Cambiar</button>
-            <input type="hidden" name="action" value="changePassword">
-        </form>
+    <section class="section_profile">
+        <img class="img_profile" src="./assets/moai.jpg" alt="Foto de perfil">
+        <h2>Username</h2>
+    </section>
+    <h2>Tus recomendaciones</h2>
+    <section class="section_recomendations">
+        <article>
+            <h3>Ejemplo</h3>
+            <p>Texto</p>
+        </article>
+        <article>
+            <h3>One Piece</h3>
+            <p>One Piece es una obra increíble que combina aventura, humor y emociones profundas. La historia de Luffy y su tripulación está llena de momentos épicos y personajes memorables. El desarrollo del mundo y los arcos argumentales mantienen al espectador enganchado, y el mensaje sobre la amistad y perseguir los sueños es inspirador.</p>
+        </article>
+        <article>
+            <img class="img_publication" src="./assets/moai.jpg" alt="Foto de la publicacion">
+            <h3>Ejemplo</h3>
+            <p>Texto</p>
+        </article>
+        <article></article>
+        <article></article>
     </section>
 </body>
 </html>
